Address review findings: security, correctness, and cleanup
Security:
- ABAC action on record_event: "read" → "update" (write op needs write check)
- Scope parameter: free-form string → Literal["cradle-to-gate", ...]
- Engine rejects unknown scopes with ValueError instead of silent fallback
- Add field bounds: MaterialItem.mass_kg >= 0, quantity >= 1, revisions >= 1

Correctness:
- Factor loader: wrap per-entry parsing in try/except (corrupt entry no longer
  kills entire database load)
- Float → Double for GWP column (64-bit precision for scientific calculations)
- Handlers: check digital_thread_auto_record setting, use savepoint to avoid
  poisoning parent transaction on failure

Cleanup:
- Remove dead TimelinePhaseGroup schema
- Remove redundant COUNT query in projections (use len(rows))
- Fix import consistency: security.abac → security re-export
- Remove late import of LifecyclePhase in router
- Add missing "from exc" on ValueError re-raise
- Fix _get_dpp_or_404 return type: object → DPP

+++ backend/app/db/migrations/versions/0014_digital_thread_lca_tables.py
@@ -163,7 +163,7 @@ def upgrade() -> None:
         ),
         sa.Column(
             "total_gwp_kg_co2e",
-            sa.Float(),
+            sa.Double(),
             nullable=False,
             comment="Total GWP in kg CO2 equivalent",
         ),
